const-oid: factor traits into module
Adds a `traits` module containing `AssociatedOid` and `DynAssociatedOid`
